requirements.txt
SQLAlchemy==1.0.5
biopython==1.65
blosc==1.2.8